Переносит метод parseFormData в контроллер задачи
